requirements.txt
opencv
Pillow
numpy